move interface and model to seperate file--core
