+++ Core/Models/Model.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace SPA_Angular.NETCore.Models
+namespace SPA_Angular.NETCore.Core.Models
 {
     [Table("Models")]
     public class Model
