+++ Persistence/SpaDbContext.cs
@@ -1,5 +1,5 @@
 using Microsoft.EntityFrameworkCore;
-using SPA_Angular.NETCore.Models;
+using SPA_Angular.NETCore.Core.Models;
 
 namespace SPA_Angular.NETCore.Persistence
 {
